test main method thinking
I want my code to be run from a single program and all the variables that are to be sent between are to be predefined and created from previous iterations.

could also be good to have all programs call and write to some sort of "components" or static word document so all changes across the document can be handled there.

atm all of the stuff will be manually configured.